add admin tests
